added simple dictionary attack

+++ HashDog/BruteForceGenerator.cpp
@@ -2,7 +2,7 @@
 #include "pch.h"
 
 
-BruteForceGenerator::BruteForceGenerator(int password_size) {
+BruteForceGenerator::BruteForceGenerator(size_t password_size) {
 	length = password_size;
 	internal_pointers = new std::bitset<pointer_size>[password_size];
 }
@@ -12,7 +12,6 @@ BruteForceGenerator::~BruteForceGenerator() {
 }
 
 void BruteForceGenerator::increment_pointer(int index) {
-	
 	for (size_t i = 0; i < pointer_size; ++i) {
 		if (internal_pointers[index][i] == 0) {
 			for (int j = 0; j <= i; ++j) {
@@ -28,11 +27,11 @@ void BruteForceGenerator::increment_pointer(int index) {
 }
 
 void BruteForceGenerator::set_password_candidate(char *candidate) {
-	for (int i = 0; i < length; ++i) {
+	for (size_t i = 0; i < length; ++i) {
 		candidate[i] = letters[(uint8_t)internal_pointers[i].to_ulong()];
 	}
 	candidate[length] = '\0';
-	increment_pointer(length - 1);
+	increment_pointer((int)length - 1);
 }
 
 void BruteForceGenerator::set_start_value(int value, int pointer_index) {
